@@ -20,6 +20,7 @@
     "@types/express": "4.17.6",
     "@types/flat": "5.0.0",
     "@types/geojson": "7946.0.7",
+    "@types/git-revision-webpack-plugin": "3.0.0",
     "@types/hapi__joi": "15.0.4",
     "@types/jsonwebtoken": "8.3.9",
     "@types/jwt-decode": "2.2.1",
@@ -27,59 +28,57 @@
     "@types/mockdate": "2.0.0",
     "@types/moment-timezone": "0.5.13",
     "@types/morgan": "1.9.0",
-    "@types/node": "12.12.24",
+    "@types/node": "12.12.37",
     "@types/node-rsa": "1.0.0",
     "@types/pg": "7.14.3",
     "@types/point-in-polygon": "1.0.0",
-    "@types/redis": "2.8.17",
+    "@types/redis": "2.8.18",
     "@types/request-promise": "4.1.46",
     "@types/sinon": "9.0.0",
     "@types/sinon-express-mock": "1.3.8",
     "@types/supertest": "2.0.8",
     "@types/uuid": "7.0.2",
-    "@types/ws": "7.2.3",
+    "@types/webpack": "4.41.12",
+    "@types/webpack-merge": "4.1.5",
+    "@types/ws": "7.2.4",
     "@types/yargs": "15.0.4",
-    "@typescript-eslint/eslint-plugin": "2.27.0",
-    "@typescript-eslint/parser": "2.27.0",
+    "@typescript-eslint/eslint-plugin": "2.29.0",
+    "@typescript-eslint/parser": "2.29.0",
     "chai": "4.2.0",
     "check-node-version": "4.0.3",
     "cypress": "4.1.0",
     "dotenv": "8.2.0",
     "eslint": "6.8.0",
     "eslint-config-airbnb-base": "14.1.0",
-    "eslint-config-prettier": "6.10.1",
+    "eslint-config-prettier": "6.11.0",
     "eslint-config-recommended": "4.0.0",
     "eslint-plugin-import": "2.20.2",
-    "eslint-plugin-prettier": "3.1.2",
+    "eslint-plugin-prettier": "3.1.3",
     "eslint-plugin-promise": "4.2.1",
-    "git-revision-webpack-plugin": "3.0.4",
     "husky": "4.2.5",
     "lerna": "3.20.2",
     "mocha": "7.1.1",
-    "mochawesome": "6.0.0",
+    "mochawesome": "6.1.0",
     "mockdate": "2.0.5",
     "node-rsa": "1.0.8",
     "nodemon": "2.0.3",
     "nyc": "15.0.1",
-    "prettier": "2.0.4",
+    "prettier": "2.0.5",
     "should": "13.2.3",
     "sinon": "9.0.2",
     "sinon-express-mock": "2.2.1",
-    "source-map-support": "0.5.16",
+    "source-map-support": "0.5.18",
     "supertest": "4.0.2",
-    "ts-loader": "6.2.2",
+    "ts-loader": "7.0.1",
     "ts-mocha": "7.0.0",
-    "ts-node": "8.8.2",
+    "ts-node": "8.9.0",
     "tsconfig-paths": "3.9.0",
     "typescript": "3.8.3",
     "unit.js": "2.1.1",
     "uuid": "7.0.3",
-    "webpack": "4.42.1",
-    "webpack-bundle-analyzer": "3.6.1",
-    "webpack-cli": "3.3.11",
-    "zip-webpack-plugin": "3.0.0"
+    "webpack": "4.43.0",
+    "webpack-cli": "3.3.11"
   },
-  "types": "types/mds.d.ts",
   "scripts": {
     "build": "yarn run:concurrent build",
     "bundle": "yarn run:concurrent bundle",
