@@ -15,12 +15,12 @@
  */
 
 import { VehicleTelemetryProcessor } from '@mds-core/mds-stream-processor'
-import { env } from '@container-images/env-inject'
 import logger from '@mds-core/mds-logger'
 
-const { npm_package_name, npm_package_version, npm_package_git_commit, KAFKA_HOST } = env()
+const { npm_package_name, npm_package_version, npm_package_git_commit, KAFKA_HOST } = process.env
 
-VehicleTelemetryProcessor.start()
+VehicleTelemetryProcessor()
+  .start()
   .then(() => {
     logger.info(
       `Running ${npm_package_name} v${npm_package_version} (${
