@@ -19,7 +19,7 @@ import {
   UUID
 } from '@mds-core/mds-types'
 import urls from 'url'
-import { parseQuery } from '@mds-core/mds-api-helpers'
+import { parseRequest } from '@mds-core/mds-api-helpers'
 import {
   badDevice,
   getVehicles,
@@ -104,7 +104,7 @@ export const getVehicleById = async (req: AgencyApiRequest, res: AgencyApiRespon
   const { device_id } = req.params
 
   const { provider_id } = res.locals.scopes.includes('vehicles:read')
-    ? parseQuery(req.query).keys('provider_id')
+    ? parseRequest(req).query('provider_id')
     : res.locals
 
   const payload = await readPayload(device_id)
@@ -122,7 +122,7 @@ export const getVehicleById = async (req: AgencyApiRequest, res: AgencyApiRespon
 export const getVehiclesByProvider = async (req: AgencyApiRequest, res: AgencyApiResponse) => {
   const PAGE_SIZE = 1000
 
-  const { skip = 0, take = PAGE_SIZE } = parseQuery(req.query, Number).keys('skip', 'take')
+  const { skip = 0, take = PAGE_SIZE } = parseRequest(req, { parser: Number }).query('skip', 'take')
 
   const url = urls.format({
     protocol: req.get('x-forwarded-proto') || req.protocol,
@@ -132,7 +132,7 @@ export const getVehiclesByProvider = async (req: AgencyApiRequest, res: AgencyAp
 
   // TODO: Replace with express middleware
   const { provider_id } = res.locals.scopes.includes('vehicles:read')
-    ? parseQuery(req.query).keys('provider_id')
+    ? parseRequest(req).query('provider_id')
     : res.locals
 
   try {
