@@ -17,4 +17,4 @@
 import { HttpServer, ApiServer } from '@mds-core/mds-api-server'
 import { api } from './api'
 
-HttpServer(process.env.PORT ?? 4001, ApiServer(api))
+HttpServer(ApiServer(api), { port: process.env.AGENCY_API_PORT })
