@@ -1,15 +1,17 @@
 import { UUID, Device, VehicleEvent, Telemetry } from '@mds-core/mds-types'
 import { MultiPolygon } from 'geojson'
-import { ApiRequest, ApiResponse, ApiResponseLocals } from '@mds-core/mds-api-server'
+import { ApiRequest, ApiResponse, ApiClaims } from '@mds-core/mds-api-server'
 import { Params, ParamsDictionary } from 'express-serve-static-core'
 
 export type AgencyApiRequest<P extends Params = ParamsDictionary> = ApiRequest<P>
 
-export interface AgencyApiResponse extends ApiResponse {
-  locals: ApiResponseLocals & {
+export type AgencyApiAccessTokenScopes = 'admin:all' | 'vehicles:read'
+
+export type AgencyApiResponse = ApiResponse<
+  ApiClaims<AgencyApiAccessTokenScopes> & {
     provider_id: UUID
   }
-}
+>
 
 export interface ServiceArea {
   service_area_id: UUID
