@@ -16,22 +16,11 @@
 
 import urls from 'url'
 import express from 'express'
+import { parseObjectProperties, ParseObjectPropertiesOptions } from '@mds-core/mds-utils'
 
 interface PagingParams {
   skip: number
   take: number
-}
-
-export const asPagingParams: <T extends Partial<{ [P in keyof PagingParams]: unknown }>>(
-  params: T
-) => T & PagingParams = params => {
-  const [DEFAULT_PAGE_SIZE, MAX_PAGE_SIZE] = [100, 1000]
-  const [skip, take] = [params.skip, params.take].map(Number)
-  return {
-    ...params,
-    skip: Number.isNaN(skip) || skip <= 0 ? 0 : skip,
-    take: Number.isNaN(take) || take <= 0 ? DEFAULT_PAGE_SIZE : Math.min(take, MAX_PAGE_SIZE)
-  }
 }
 
 const jsonApiLink = (req: express.Request, skip: number, take: number): string =>
@@ -55,12 +44,17 @@ export const asJsonApiLinks = (req: express.Request, skip: number, take: number,
   return undefined
 }
 
-export const parseQuery = <T = string>(query: { [k: string]: unknown }, parser?: (val: string) => T) => {
+export const parseRequest = <T = string>(req: express.Request, options?: ParseObjectPropertiesOptions<T>) => {
+  const { keys: query } = parseObjectProperties<T>(req.query, options)
+  const { keys: params } = parseObjectProperties<T>(req.params, options)
+  return { params, query }
+}
+
+export const parsePagingQueryParams = (req: express.Request) => {
+  const [DEFAULT_PAGE_SIZE, MAX_PAGE_SIZE] = [100, 1000]
+  const { skip = 0, take = DEFAULT_PAGE_SIZE } = parseRequest(req, { parser: Number }).query('skip', 'take')
   return {
-    keys: <TKey extends string>(first: TKey, ...rest: TKey[]): Partial<{ [P in TKey]: T }> =>
-      [first, ...rest]
-        .map(key => ({ key, value: query[key] }))
-        .filter((param): param is { key: TKey; value: string } => typeof param.value === 'string')
-        .reduce((params, { key, value }) => ({ ...params, [key]: parser ? parser(value) : value }), {})
+    skip: Number.isNaN(skip) ? 0 : Math.max(0, skip),
+    take: Number.isNaN(take) ? DEFAULT_PAGE_SIZE : Math.min(take, MAX_PAGE_SIZE)
   }
 }
