@@ -55,8 +55,8 @@ import {
   TelemetryData,
   VEHICLE_EVENT
 } from '@mds-core/mds-types'
-import { asPagingParams, asJsonApiLinks, parseQuery } from '@mds-core/mds-api-helpers'
-import { checkAccess } from '@mds-core/mds-api-server'
+import { parsePagingQueryParams, asJsonApiLinks, parseRequest } from '@mds-core/mds-api-helpers'
+import { checkAccess, AccessTokenScopeValidator } from '@mds-core/mds-api-server'
 import {
   AuditApiAuditEndRequest,
   AuditApiAuditNoteRequest,
@@ -68,7 +68,8 @@ import {
   AuditApiResponse,
   AuditApiTripRequest,
   AuditApiVehicleEventRequest,
-  AuditApiVehicleTelemetryRequest
+  AuditApiVehicleTelemetryRequest,
+  AuditApiAccessTokenScopes
 } from './types'
 import {
   deleteAudit,
@@ -110,6 +111,8 @@ function flattenTelemetry(telemetry?: Telemetry): TelemetryData {
         charge: null
       }
 }
+
+const checkAuditApiAccess = (validator: AccessTokenScopeValidator<AuditApiAccessTokenScopes>) => checkAccess(validator)
 
 function api(app: express.Express): express.Express {
   /**
@@ -168,7 +171,7 @@ function api(app: express.Express): express.Express {
    */
   app.post(
     pathsFor('/trips/:audit_trip_id/start'),
-    checkAccess(scopes => scopes.includes('audits:write')),
+    checkAuditApiAccess(scopes => scopes.includes('audits:write')),
     async (req: AuditApiAuditStartRequest, res: AuditApiResponse) => {
       try {
         const { audit_trip_id, audit, audit_subject_id, recorded } = res.locals
@@ -252,7 +255,7 @@ function api(app: express.Express): express.Express {
    */
   app.post(
     pathsFor('/trips/:audit_trip_id/vehicle/event'),
-    checkAccess(scopes => scopes.includes('audits:write')),
+    checkAuditApiAccess(scopes => scopes.includes('audits:write')),
     async (req: AuditApiVehicleEventRequest, res: AuditApiResponse) => {
       try {
         const { audit_trip_id, audit_subject_id, audit, recorded } = res.locals
@@ -302,7 +305,7 @@ function api(app: express.Express): express.Express {
    */
   app.post(
     pathsFor('/trips/:audit_trip_id/vehicle/telemetry'),
-    checkAccess(scopes => scopes.includes('audits:write')),
+    checkAuditApiAccess(scopes => scopes.includes('audits:write')),
     async (req: AuditApiVehicleTelemetryRequest, res: AuditApiResponse) => {
       try {
         const { audit_trip_id, audit_subject_id, audit, recorded } = res.locals
@@ -348,7 +351,7 @@ function api(app: express.Express): express.Express {
    */
   app.post(
     [...pathsFor('/trips/:audit_trip_id/note'), ...pathsFor('/trips/:audit_trip_id/event')],
-    checkAccess(scopes => scopes.includes('audits:write')),
+    checkAuditApiAccess(scopes => scopes.includes('audits:write')),
     async (req: AuditApiAuditNoteRequest, res: AuditApiResponse) => {
       try {
         const { audit_trip_id, audit, audit_subject_id, recorded } = res.locals
@@ -414,7 +417,7 @@ function api(app: express.Express): express.Express {
    */
   app.post(
     pathsFor('/trips/:audit_trip_id/end'),
-    checkAccess(scopes => scopes.includes('audits:write')),
+    checkAuditApiAccess(scopes => scopes.includes('audits:write')),
     async (req: AuditApiAuditEndRequest, res: AuditApiResponse) => {
       try {
         const { audit_trip_id, audit, audit_subject_id, recorded } = res.locals
@@ -464,7 +467,7 @@ function api(app: express.Express): express.Express {
    */
   app.get(
     pathsFor('/trips/:audit_trip_id'),
-    checkAccess(scopes => scopes.includes('audits:read')),
+    checkAuditApiAccess(scopes => scopes.includes('audits:read')),
     async (req: AuditApiGetTripRequest, res: AuditApiResponse<AuditDetails>) => {
       try {
         const { audit_trip_id, audit } = res.locals
@@ -511,9 +514,9 @@ function api(app: express.Express): express.Express {
                 }
               }, {})
 
-            const { event_viewport_adjustment = seconds(30) } = parseQuery(req.query, x => seconds(Number(x))).keys(
-              'event_viewport_adjustment'
-            )
+            const { event_viewport_adjustment = seconds(30) } = parseRequest(req, {
+              parser: x => seconds(Number(x))
+            }).query('event_viewport_adjustment')
 
             const start_time = audit_start && audit_start - event_viewport_adjustment
             const end_time = (end => end && end + event_viewport_adjustment)(audit_end || last_event)
@@ -575,7 +578,7 @@ function api(app: express.Express): express.Express {
 
   app.delete(
     pathsFor('/trips/:audit_trip_id'),
-    checkAccess(scopes => scopes.includes('audits:delete')),
+    checkAuditApiAccess(scopes => scopes.includes('audits:delete')),
     async (req: AuditApiTripRequest, res: AuditApiResponse) => {
       try {
         const { audit_trip_id, audit } = res.locals
@@ -604,19 +607,16 @@ function api(app: express.Express): express.Express {
    */
   app.get(
     pathsFor('/trips'),
-    checkAccess(scopes => scopes.includes('audits:read')),
+    checkAuditApiAccess(scopes => scopes.includes('audits:read')),
     async (req: AuditApiGetTripsRequest, res: AuditApiResponse) => {
       try {
-        const { start_time, end_time } = req.query
-        const { skip, take } = asPagingParams(req.query)
+        const { skip, take } = parsePagingQueryParams(req)
 
-        // Construct the query params
         const query = {
-          ...req.query,
+          ...parseRequest(req, { parser: Number }).query('start_time', 'end_time'),
+          ...parseRequest(req).query('provider_id', 'provider_vehicle_id', 'audit_subject_id'),
           skip,
-          take,
-          start_time: start_time ? Number(start_time) : undefined,
-          end_time: end_time ? Number(end_time) : undefined
+          take
         }
 
         // Query the audits
@@ -648,12 +648,12 @@ function api(app: express.Express): express.Express {
    */
   app.get(
     pathsFor('/vehicles'),
-    checkAccess(scopes => scopes.includes('audits:vehicles:read')),
+    checkAuditApiAccess(scopes => scopes.includes('audits:vehicles:read')),
     async (req, res) => {
       const { skip, take } = { skip: 0, take: 10000 }
       const { strict = true, bbox, provider_id } = {
-        ...parseQuery(req.query, JSON.parse).keys('strict', 'bbox'),
-        ...parseQuery(req.query).keys('provider_id')
+        ...parseRequest(req, { parser: JSON.parse }).query('strict', 'bbox'),
+        ...parseRequest(req).query('provider_id')
       }
 
       const url = urls.format({
