@@ -1,10 +1,13 @@
 import { UUID } from '@mds-core/mds-types'
 import { MultiPolygon } from 'geojson'
-import { ApiRequest, ApiResponse } from '@mds-core/mds-api-server'
+import { ApiRequest, ApiResponse, ApiClaims } from '@mds-core/mds-api-server'
 import { Params, ParamsDictionary } from 'express-serve-static-core'
 
 export type DailyApiRequest<P extends Params = ParamsDictionary> = ApiRequest<P>
-export type DailyApiResponse = ApiResponse
+
+export type DailyApiAccessTokenScopes = 'admin:all'
+
+export type DailyApiResponse = ApiResponse<ApiClaims<DailyApiAccessTokenScopes>>
 
 export interface ServiceArea {
   service_area_id: UUID
