@@ -6,11 +6,9 @@ import {
   TelemetryData,
   VEHICLE_TYPE,
   PROPULSION_TYPE,
-  PROVIDER_EVENT,
-  PROVIDER_REASON,
   Nullable
 } from '@mds-core/mds-types'
-import { Feature, FeatureCollection } from 'geojson'
+import { FeatureCollection } from 'geojson'
 
 export interface ReadEventsResult {
   events: Recorded<VehicleEvent>[]
@@ -38,24 +36,6 @@ export interface Trip {
   actual_cost?: number | null
   recorded: Timestamp
 }
-
-export interface StatusChange {
-  provider_id: UUID
-  provider_name: string
-  device_id: UUID
-  vehicle_id: string
-  vehicle_type: VEHICLE_TYPE
-  propulsion_type: PROPULSION_TYPE[]
-  event_type: PROVIDER_EVENT
-  event_type_reason: PROVIDER_REASON
-  event_time: Timestamp
-  event_location: Feature | null
-  battery_pct: number | null
-  associated_trip: UUID | null
-  recorded: Timestamp
-}
-
-export type StatusChangeEvent = Pick<StatusChange, 'event_type' | 'event_type_reason'>
 
 // Represents a row in the "telemetry" table
 export interface TelemetryRecord extends TelemetryData {
