@@ -4,20 +4,24 @@ import db from '@mds-core/mds-db'
 import { pathsFor, ServerError, NotFoundError, InsufficientPermissionsError, BadParamsError } from '@mds-core/mds-utils'
 import logger from '@mds-core/mds-logger'
 
-import { checkAccess } from '@mds-core/mds-api-server'
-import { parseQuery } from '@mds-core/mds-api-helpers'
+import { checkAccess, AccessTokenScopeValidator } from '@mds-core/mds-api-server'
+import { parseRequest } from '@mds-core/mds-api-helpers'
+import { GeographyApiAccessTokenScopes, GeographyApiResponse, GeographyApiRequest } from './types'
+
+const checkGeographyApiAccess = (validator: AccessTokenScopeValidator<GeographyApiAccessTokenScopes>) =>
+  checkAccess(validator)
 
 function api(app: express.Express): express.Express {
   app.get(
     pathsFor('/geographies/meta/'),
-    checkAccess(scopes => {
+    checkGeographyApiAccess(scopes => {
       return scopes.includes('geographies:read:published') || scopes.includes('geographies:read:unpublished')
     }),
-    async (req, res) => {
+    async (req: GeographyApiRequest, res: GeographyApiResponse) => {
       const { scopes } = res.locals
       const params = {
         ...{ get_published: null, get_unpublished: null },
-        ...parseQuery(req.query, x => (x ? x === 'true' : null)).keys('get_published', 'get_unpublished')
+        ...parseRequest(req, { parser: x => (x ? x === 'true' : null) }).query('get_published', 'get_unpublished')
       }
 
       /* If the user can only read published geos, and all they want is the unpublished metadata,
@@ -64,10 +68,10 @@ function api(app: express.Express): express.Express {
 
   app.get(
     pathsFor('/geographies/:geography_id'),
-    checkAccess(scopes => {
+    checkGeographyApiAccess(scopes => {
       return scopes.includes('geographies:read:published') || scopes.includes('geographies:read:unpublished')
     }),
-    async (req, res) => {
+    async (req: GeographyApiRequest, res: GeographyApiResponse) => {
       const { geography_id } = req.params
       try {
         const geography = await db.readSingleGeography(geography_id)
@@ -92,10 +96,10 @@ function api(app: express.Express): express.Express {
 
   app.get(
     pathsFor('/geographies'),
-    checkAccess(scopes => {
+    checkGeographyApiAccess(scopes => {
       return scopes.includes('geographies:read:published') || scopes.includes('geographies:read:unpublished')
     }),
-    async (req, res) => {
+    async (req: GeographyApiRequest, res: GeographyApiResponse) => {
       const summary = req.query.summary === 'true'
       const { get_published, get_unpublished } = req.query
       const params = {
@@ -131,10 +135,10 @@ function api(app: express.Express): express.Express {
 
   app.get(
     pathsFor('/geographies/:geography_id/meta'),
-    checkAccess(scopes => {
+    checkGeographyApiAccess(scopes => {
       return scopes.includes('geographies:read:published') || scopes.includes('geographies:read:unpublished')
     }),
-    async (req, res) => {
+    async (req: GeographyApiRequest, res: GeographyApiResponse) => {
       const { geography_id } = req.params
       try {
         const geography_metadata = await db.readSingleGeographyMetadata(geography_id)
