@@ -16,41 +16,49 @@
 
 import express from 'express'
 import { pathsFor } from '@mds-core/mds-utils'
-import { checkAccess } from '@mds-core/mds-api-server'
+import { checkAccess, AccessTokenScopeValidator } from '@mds-core/mds-api-server'
 import { JurisdictionApiVersionMiddleware } from './middleware'
 import {
   CreateJurisdictionHandler,
   DeleteJurisdictionHandler,
-  GetAllJurisdictionsHandler,
-  GetOneJurisdictionHandler,
+  GetJurisdictionsHandler,
+  GetJurisdictionHandler,
   UpdateJurisdictionHandler
 } from './handlers'
+import { JurisdictionApiAccessTokenScopes } from './types'
+
+const checkJurisdictionApiAccess = (validator: AccessTokenScopeValidator<JurisdictionApiAccessTokenScopes>) =>
+  checkAccess(validator)
 
 export const api = (app: express.Express): express.Express =>
   app
     .use(JurisdictionApiVersionMiddleware)
     .get(
       pathsFor('/jurisdictions'),
-      checkAccess(scopes => scopes.includes('jurisdictions:read') || scopes.includes('jurisdictions:read:claim')),
-      GetAllJurisdictionsHandler
+      checkJurisdictionApiAccess(
+        scopes => scopes.includes('jurisdictions:read') || scopes.includes('jurisdictions:read:claim')
+      ),
+      GetJurisdictionsHandler
     )
     .get(
       pathsFor('/jurisdictions/:jurisdiction_id'),
-      checkAccess(scopes => scopes.includes('jurisdictions:read') || scopes.includes('jurisdictions:read:claim')),
-      GetOneJurisdictionHandler
+      checkJurisdictionApiAccess(
+        scopes => scopes.includes('jurisdictions:read') || scopes.includes('jurisdictions:read:claim')
+      ),
+      GetJurisdictionHandler
     )
     .post(
       pathsFor('/jurisdictions'),
-      checkAccess(scopes => scopes.includes('jurisdictions:write')),
+      checkJurisdictionApiAccess(scopes => scopes.includes('jurisdictions:write')),
       CreateJurisdictionHandler
     )
     .put(
       pathsFor('/jurisdictions/:jurisdiction_id'),
-      checkAccess(scopes => scopes.includes('jurisdictions:write')),
+      checkJurisdictionApiAccess(scopes => scopes.includes('jurisdictions:write')),
       UpdateJurisdictionHandler
     )
     .delete(
       pathsFor('/jurisdictions/:jurisdiction_id'),
-      checkAccess(scopes => scopes.includes('jurisdictions:write')),
+      checkJurisdictionApiAccess(scopes => scopes.includes('jurisdictions:write')),
       DeleteJurisdictionHandler
     )
