@@ -19,8 +19,8 @@ import {
   CreateJurisdictionType,
   JurisdictionDomainModel
 } from '@mds-core/mds-jurisdiction-service'
+import { HandleServiceResponse } from '@mds-core/mds-service-helpers'
 import { JurisdictionApiRequest, JurisdictionApiResponse } from '../types'
-import { UnexpectedServiceError } from './utils'
 
 interface CreateJurisdictionRequest extends JurisdictionApiRequest {
   body: CreateJurisdictionType | CreateJurisdictionType[]
@@ -36,29 +36,24 @@ type CreateJurisdictionResponse = JurisdictionApiResponse<
 >
 
 export const CreateJurisdictionHandler = async (req: CreateJurisdictionRequest, res: CreateJurisdictionResponse) => {
-  const [error, jurisdictions] = await JurisdictionServiceClient.createJurisdictions(
-    Array.isArray(req.body) ? req.body : [req.body]
+  HandleServiceResponse(
+    await JurisdictionServiceClient.createJurisdictions(Array.isArray(req.body) ? req.body : [req.body]),
+    error => {
+      if (error.type === 'ValidationError') {
+        return res.status(400).send({ error })
+      }
+      if (error.type === 'ConflictError') {
+        return res.status(409).send({ error })
+      }
+      return res.status(500).send({ error })
+    },
+    jurisdictions => {
+      const { version } = res.locals
+      if (!Array.isArray(req.body)) {
+        const [jurisdiction] = jurisdictions
+        return res.status(201).send({ version, jurisdiction })
+      }
+      return res.status(201).send({ version, jurisdictions })
+    }
   )
-
-  // Handle result
-  if (jurisdictions) {
-    if (!Array.isArray(req.body)) {
-      const [jurisdiction] = jurisdictions
-      return res.status(201).send({ version: res.locals.version, jurisdiction })
-    }
-    return res.status(201).send({ version: res.locals.version, jurisdictions })
-  }
-
-  // Handle errors
-  if (error) {
-    if (error.type === 'ValidationError') {
-      return res.status(400).send({ error })
-    }
-    if (error.type === 'ConflictError') {
-      return res.status(409).send({ error })
-    }
-    return res.status(500).send({ error })
-  }
-
-  return res.status(500).send(UnexpectedServiceError)
 }
