@@ -20,8 +20,8 @@ import {
   JurisdictionDomainModel
 } from '@mds-core/mds-jurisdiction-service'
 import { UUID } from '@mds-core/mds-types'
+import { HandleServiceResponse } from '@mds-core/mds-service-helpers'
 import { JurisdictionApiRequest, JurisdictionApiResponse } from '../types'
-import { UnexpectedServiceError } from './utils'
 
 interface UpdateJurisdictionRequest extends JurisdictionApiRequest<{ jurisdiction_id: UUID }> {
   body: UpdateJurisdictionType
@@ -32,26 +32,24 @@ type UpdateJurisdictionResponse = JurisdictionApiResponse<{
 }>
 
 export const UpdateJurisdictionHandler = async (req: UpdateJurisdictionRequest, res: UpdateJurisdictionResponse) => {
-  const [error, jurisdiction] = await JurisdictionServiceClient.updateJurisdiction(req.params.jurisdiction_id, req.body)
-
-  // Handle result
-  if (jurisdiction) {
-    return res.status(200).send({ version: res.locals.version, jurisdiction })
-  }
-
-  // Handle errors
-  if (error) {
-    if (error.type === 'ValidationError') {
-      return res.status(400).send({ error })
+  const { jurisdiction_id } = req.params
+  HandleServiceResponse(
+    await JurisdictionServiceClient.updateJurisdiction(jurisdiction_id, req.body),
+    error => {
+      if (error.type === 'ValidationError') {
+        return res.status(400).send({ error })
+      }
+      if (error.type === 'NotFoundError') {
+        return res.status(404).send({ error })
+      }
+      if (error.type === 'ConflictError') {
+        return res.status(409).send({ error })
+      }
+      return res.status(500).send({ error })
+    },
+    jurisdiction => {
+      const { version } = res.locals
+      return res.status(200).send({ version, jurisdiction })
     }
-    if (error.type === 'NotFoundError') {
-      return res.status(404).send({ error })
-    }
-    if (error.type === 'ConflictError') {
-      return res.status(409).send({ error })
-    }
-    return res.status(500).send({ error })
-  }
-
-  return res.status(500).send(UnexpectedServiceError)
+  )
 }
