@@ -1,4 +1,11 @@
-import { ApiRequest, ApiResponse } from '@mds-core/mds-api-server'
+import { ApiRequest, ApiResponse, ApiClaims } from '@mds-core/mds-api-server'
 
-export type PolicyApiRequest = ApiRequest
-export type PolicyApiResponse = ApiResponse
+export type PolicyAuthorApiRequest = ApiRequest
+
+export type PolicyAuthorApiAccessTokenScopes =
+  | 'policies:read'
+  | 'policies:write'
+  | 'policies:publish'
+  | 'policies:delete'
+
+export type PolicyAuthorApiResponse = ApiResponse<ApiClaims<PolicyAuthorApiAccessTokenScopes>>
