@@ -1,4 +1,7 @@
-import { ApiRequest, ApiResponse } from '@mds-core/mds-api-server'
+import { ApiRequest, ApiResponse, ApiClaims } from '@mds-core/mds-api-server'
 
 export type PolicyApiRequest = ApiRequest
-export type PolicyApiResponse = ApiResponse
+
+export type PolicyApiAccessTokenScopes = never
+
+export type PolicyApiResponse = ApiResponse<ApiClaims<PolicyApiAccessTokenScopes>>
