@@ -10,10 +10,7 @@ import { Clients } from './clients'
 import { ENTITY_TYPE } from './types'
 
 export const WebSocketServer = async () => {
-  const server = HttpServer(
-    process.env.PORT ?? 4009,
-    ApiServer(app => app)
-  )
+  const server = HttpServer(ApiServer(app => app))
 
   logger.info('Creating WS server')
   const wss = new WebSocket.Server({ server })
